Add ability to subscribe to feeds via --subscribe flag

+++ src/romanpkg/subscriptions.nim
@@ -7,12 +7,17 @@ type
     url*: string
 
 
+proc getSubsFilePath(): string {.raises: [].} =
+  joinPath(getConfigDir(), "roman", "subscriptions")
+
+
+
 proc initConfigDir() {.raises: [RomanError].} =
   let configDir = joinPath(getConfigDir(), "roman")
   try:
-    discard existsOrCreateDir(configDir)
-    let subsFile = joinPath(configDir, "subscriptions")
-    writeFile(subsFile, "")
+    if not existsOrCreateDir(configDir):
+      let subsFile = joinPath(configDir, "subscriptions")
+      writeFile(subsFile, "")
   except OSError as e:
     raise newException(RomanError, e.msg)
   except IOError as e:
@@ -20,12 +25,12 @@ proc initConfigDir() {.raises: [RomanError].} =
 
 
 proc readConfigFile(): seq[string] {.raises: [RomanError].} =
-  let configFile = joinPath(getConfigDir(), "roman", "subscriptions")
-  if not existsFile(configFile):
+  let subsFilePath = getSubsFilePath()
+  if not existsFile(subsFilePath):
     initConfigDir()
     return
   try:
-    for line in lines(configFile):
+    for line in lines(subsFilePath):
       result.add(line)
   except IOError as e:
     raise newException(RomanError, e.msg)
@@ -34,3 +39,13 @@ proc readConfigFile(): seq[string] {.raises: [RomanError].} =
 proc getSubscriptions*(): seq[Subscription] {.raises: [RomanError].} =
   for url in readConfigFile():
     result.add(Subscription(url: url))
+
+
+proc addSubscriptionToSubsFile*(url: string) {.raises: [RomanError].} =
+  var f: File
+  let filename = getSubsFilePath()
+  try:
+    if f.open(filename, fmAppend):
+      f.writeLine(url)
+  except IOError as e:
+    raise newException(RomanError, e.msg)
